Revert "Squelette + Architecture"
