--- app/src/main/java/com/project/remmeds/remmeds/Configuration.java
@@ -1,13 +1,0 @@
-package com.project.remmeds.remmeds;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class Configuration extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_configuration);
-    }
-}
