--- app/src/main/res/layout/activity_ajout_contact.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".AjoutContact">
-
-    <EditText
-        android:id="@+id/editTextPrenom"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_alignStart="@+id/editTextAdresse"
-        android:layout_marginTop="194dp"
-        android:ems="10"
-        android:hint="@string/prenom_activity_ajout_contact"
-        android:inputType="text" />
-
-    <EditText
-        android:id="@+id/editTextNom"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_alignStart="@+id/editTextAdresse"
-        android:layout_marginTop="115dp"
-        android:ems="10"
-        android:hint="@string/nom_activity_ajout_contact"
-        android:inputType="text" />
-
-    <EditText
-        android:id="@+id/editTextNumero"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_alignStart="@+id/editTextAdresse"
-        android:layout_marginBottom="200dp"
-        android:ems="10"
-        android:hint="@string/numero_activity_ajout_contact"
-        android:inputType="number" />
-
-    <EditText
-        android:id="@+id/editTextAdresse"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignEnd="@+id/textView5"
-        android:layout_alignParentBottom="true"
-        android:layout_marginBottom="118dp"
-        android:ems="10"
-        android:hint="@string/adresse_activity_ajout_contact"
-        android:inputType="text" />
-
-    <TextView
-        android:id="@+id/textView5"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="47dp"
-        android:text="@string/title_activity_ajout_contact" />
-
-    <Button
-        android:id="@+id/gobutton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="36dp"
-        android:text="@string/gobutton_activity_ajout_contact" />
-</RelativeLayout>
