--- app/src/main/res/layout/content_historique.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".Historique"
-    tools:showIn="@layout/app_bar_historique">
-
-    <android.support.v7.widget.RecyclerView
-        android:layout_width="271dp"
-        android:layout_height="246dp"
-        android:layout_centerInParent="true" />
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="84dp"
-        android:layout_height="20dp"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="44dp"
-        android:text="@string/title_activity_historique" />
-
-    <Button
-        android:id="@+id/historique_view_button"
-        android:layout_width="49dp"
-        android:layout_height="36dp"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="86dp"
-        android:background="@android:drawable/ic_menu_view" />
-</RelativeLayout>
