--- app/src/main/res/layout/content_repertoire.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".Repertoire"
-    tools:showIn="@layout/app_bar_repertoire">
-
-    <Button
-        android:id="@+id/repertoire_edit_button"
-        android:layout_width="44dp"
-        android:layout_height="wrap_content"
-        android:layout_alignEnd="@+id/repertoire_add_button"
-        android:layout_alignParentBottom="true"
-        android:layout_marginBottom="75dp"
-        android:background="@android:drawable/ic_menu_edit"
-        android:layout_marginStart="53dp"/>
-
-
-    <TextView
-        android:id="@+id/text_view4"
-        android:layout_width="180dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="43dp"
-        android:text="@string/title_activity_repertoire"
-        android:textAlignment="center"
-        android:textStyle="bold"
-        tools:layout_editor_absoluteX="16dp"
-        tools:layout_editor_absoluteY="34dp" />
-
-    <android.support.v7.widget.RecyclerView
-        android:layout_width="271dp"
-        android:layout_height="246dp"
-        android:layout_centerInParent="true"
-        android:scrollbars="vertical"
-        tools:layout_editor_absoluteX="125dp"
-        tools:layout_editor_absoluteY="157dp">
-
-    </android.support.v7.widget.RecyclerView>
-
-    <Button
-        android:id="@+id/repertoire_add_button"
-        android:layout_width="27dp"
-        android:layout_height="25dp"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginEnd="53dp"
-        android:layout_marginTop="87dp"
-        android:background="@android:drawable/ic_input_add"
-        android:onClick="sendMessage" />
-
-    <Button
-        android:id="@+id/repertoire_remove_button"
-        android:layout_width="27dp"
-        android:layout_height="25dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignTop="@+id/repertoire_add_button"
-        android:layout_marginStart="53dp"
-        android:background="@android:drawable/ic_delete" />
-
-</RelativeLayout>
