+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -5,18 +5,34 @@
 
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/accueil"
-            android:icon="@drawable/ic_menu_gallery"
-            android:title="Accueil" />
-        <item
-            android:id="@+id/historique"
+            android:id="@+id/nav_camera"
             android:icon="@drawable/ic_menu_camera"
-            android:title="Historique" />
+            android:title="Import" />
         <item
-            android:id="@+id/repertoire"
+            android:id="@+id/nav_gallery"
             android:icon="@drawable/ic_menu_gallery"
-            android:title="Repertoire" />
-
+            android:title="Gallery" />
+        <item
+            android:id="@+id/nav_slideshow"
+            android:icon="@drawable/ic_menu_slideshow"
+            android:title="Slideshow" />
+        <item
+            android:id="@+id/nav_manage"
+            android:icon="@drawable/ic_menu_manage"
+            android:title="Tools" />
     </group>
+
+    <item android:title="Communicate">
+        <menu>
+            <item
+                android:id="@+id/nav_share"
+                android:icon="@drawable/ic_menu_share"
+                android:title="Share" />
+            <item
+                android:id="@+id/nav_send"
+                android:icon="@drawable/ic_menu_send"
+                android:title="Send" />
+        </menu>
+    </item>
 
 </menu>
