--- app/src/main/res/menu/activity_repertoire_drawer.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:showIn="navigation_view">
-
-    <group android:checkableBehavior="single">
-        <item
-            android:id="@+id/accueil"
-            android:icon="@drawable/ic_menu_gallery"
-            android:title="Accueil" />
-        <item
-            android:id="@+id/historique"
-            android:icon="@drawable/ic_menu_camera"
-            android:title="Historique" />
-        <item
-            android:id="@+id/repertoire"
-            android:icon="@drawable/ic_menu_gallery"
-            android:title="Repertoire" />
-    </group>
-
-</menu>
