+++ app/src/main/res/values/strings.xml
@@ -6,14 +6,4 @@
     <string name="nav_header_subtitle">android.studio@android.com</string>
     <string name="nav_header_desc">Navigation header</string>
     <string name="action_settings">Settings</string>
-    <string name="title_activity_historique">Historique</string>
-    <string name="title_activity_repertoire">Repertoire</string>
-    <string name="title_activity_main">Accueil</string>
-    <string name="title_activity_ajout_contact">Nouveau Contact</string>
-    <string name="gobutton_activity_ajout_contact">GO</string>
-    <string name="prenom_activity_ajout_contact">Prenom</string>
-    <string name="nom_activity_ajout_contact">Nom</string>
-    <string name="adresse_activity_ajout_contact">Adresse</string>
-    <string name="numero_activity_ajout_contact">Numero</string>
-    <string name="title_activity_profil">Profil</string>
 </resources>
